renamed logbuffer files to loglist
